# perf: BaseDocument improvements