--- a/frappe/model/base_document.py
+++ b/frappe/model/base_document.py
@@ -78,8 +78,9 @@ class BaseDocument(object):
 		self.update(d)
 		self.dont_update_if_missing = []
 
-		if hasattr(self, "__setup__"):
-			self.__setup__()
+		_setup_attr = getattr(self, "__setup__", None)
+		if _setup_attr:
+			_setup_attr()
 
 	@property
 	def meta(self):
@@ -101,13 +102,11 @@ class BaseDocument(object):
 				"balance": 42000
 			})
 		"""
-		if "doctype" in d:
-			self.set("doctype", d.get("doctype"))
-
 		# first set default field values of base document
 		for key in default_fields:
-			if key in d:
-				self.set(key, d.get(key))
+			_val = d.get(key)
+			if _val:
+				self.set(key, _val)
 
 		for key, value in d.items():
 			self.set(key, value)
@@ -175,20 +174,9 @@ class BaseDocument(object):
 				...
 			})
 		"""
-		if value==None:
-			value={}
-		if isinstance(value, (dict, BaseDocument)):
-			if not self.__dict__.get(key):
-				self.__dict__[key] = []
-			value = self._init_child(value, key)
-			self.__dict__[key].append(value)
-
-			# reference parent document
-			value.parent_doc = self
-
-			return value
-		else:
-
+		if value is None:
+			value = {}
+		elif not isinstance(value, (dict, BaseDocument)):
 			# metaclasses may have arbitrary lists
 			# which we can ignore
 			if (getattr(self, '_metaclass', None)
@@ -199,6 +187,15 @@ class BaseDocument(object):
 				'Document for field "{0}" attached to child table of "{1}" must be a dict or BaseDocument, not {2} ({3})'.format(key,
 					self.name, str(type(value))[1:-1], value)
 			)
+
+		self.__dict__.setdefault(key, [])
+		value = self._init_child(value, key)
+		self.__dict__[key].append(value)
+
+		# reference parent document
+		value.parent_doc = self
+
+		return value
 
 	def extend(self, key, value):
 		if isinstance(value, list):
@@ -214,12 +211,14 @@ class BaseDocument(object):
 		if not self.doctype:
 			return value
 		if not isinstance(value, BaseDocument):
-			if "doctype" not in value or value['doctype'] is None:
-				value["doctype"] = self.get_table_field_doctype(key)
-				if not value["doctype"]:
+			_doctype = value.get("doctype")
+			if not _doctype:
+				_doctype = self.get_table_field_doctype(key)
+				if not _doctype:
 					raise AttributeError(key)
+				value['doctype'] = _doctype
 
-			value = get_controller(value["doctype"])(value)
+			value = get_controller(_doctype)(value)
 			value.init_valid_columns()
 
 		value.parent = self.name