After profilling I noticed that these functions are the most expensive: `BaseDocument.__init__()`, `BaseDocument.update()`, `BaseDocument.append()`, `BaseDocument._init_child()`. After some opimizing, every request is faster up to 0.02 sec.

Update: Kept only changes in `update()` method.