@szufisher 2021-12-22
@vrslev 
would like to learn why these changes improved performance, also would like to see the before and after profiling result.

thanks.