@vrslev 2021-12-22
@szufisher Unfortunately I didn’t save profiling results. I took `frappe.app:application` WSGI app, wrapped it in Werkzeug’s profiling middleware that would save results to files.

Also I used `—profile` flag with `bench execute`.