@gavindsouza 2021-12-27
@vrslev While making performance improvements, we're going to need the benchmarking script and the numbers to back it up. Else, there's no reason to believe the improvements are due to the changes made in this PR and not attributed to the thousands of other unknowns.