@vrslev 2021-12-27
Comparing to `develop` branch:

`update()` itself is faster by ~6%

before:
```console
         22000004 function calls in 2.823 seconds

   Ordered by: standard name

   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
        1    0.000    0.000    2.823    2.823 <string>:1(<module>)
  9000000    0.974    0.000    1.306    0.000 base_document.py:157(set)
  1000000    1.243    0.000    2.617    0.000 base_document.py:95(update)
        1    0.088    0.088    2.823    2.823 t.py:18(inner)
  1000000    0.118    0.000    2.735    0.000 t.py:37(profile_update)
        1    0.000    0.000    2.823    2.823 {built-in method builtins.exec}
  9000000    0.331    0.000    0.331    0.000 {built-in method builtins.isinstance}
        1    0.000    0.000    0.000    0.000 {method 'disable' of '_lsprof.Profiler' objects}
  1000000    0.033    0.000    0.033    0.000 {method 'get' of 'dict' objects}
  1000000    0.035    0.000    0.035    0.000 {method 'items' of 'dict' objects}
```

after:

```console
         21000004 function calls in 2.667 seconds

   Ordered by: standard name

   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
        1    0.000    0.000    2.667    2.667 <string>:1(<module>)
  9000000    0.974    0.000    1.294    0.000 base_document.py:154(set)
  1000000    1.140    0.000    2.462    0.000 base_document.py:95(update)
        1    0.086    0.086    2.667    2.667 t.py:18(inner)
  1000000    0.119    0.000    2.581    0.000 t.py:37(profile_update)
        1    0.000    0.000    2.667    2.667 {built-in method builtins.exec}
  9000000    0.320    0.000    0.320    0.000 {built-in method builtins.isinstance}
        1    0.000    0.000    0.000    0.000 {method 'disable' of '_lsprof.Profiler' objects}
  1000000    0.028    0.000    0.028    0.000 {method 'items' of 'dict' objects}
```

`__init__()` is faster by ~17%.
before:
```console
         37000004 function calls in 5.080 seconds

   Ordered by: standard name

   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
        1    0.000    0.000    5.080    5.080 <string>:1(<module>)
 14000000    1.607    0.000    2.152    0.000 base_document.py:157(set)
  1000000    0.272    0.000    4.745    0.000 base_document.py:77(__init__)
  1000000    2.109    0.000    4.423    0.000 base_document.py:95(update)
        1    0.100    0.100    5.080    5.080 t.py:18(inner)
  1000000    0.234    0.000    4.980    0.000 t.py:53(profile_init)
        1    0.000    0.000    5.080    5.080 {built-in method builtins.exec}
  1000000    0.050    0.000    0.050    0.000 {built-in method builtins.hasattr}
 14000000    0.545    0.000    0.545    0.000 {built-in method builtins.isinstance}
        1    0.000    0.000    0.000    0.000 {method 'disable' of '_lsprof.Profiler' objects}
  4000000    0.121    0.000    0.121    0.000 {method 'get' of 'dict' objects}
  1000000    0.041    0.000    0.041    0.000 {method 'items' of 'dict' objects}
```

after:
```console
         31000004 function calls in 4.229 seconds

   Ordered by: standard name

   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
        1    0.000    0.000    4.229    4.229 <string>:1(<module>)
 13000000    1.436    0.000    1.923    0.000 base_document.py:154(set)
  1000000    0.267    0.000    3.906    0.000 base_document.py:77(__init__)
  1000000    1.641    0.000    3.596    0.000 base_document.py:95(update)
        1    0.096    0.096    4.229    4.229 t.py:18(inner)
  1000000    0.227    0.000    4.133    0.000 t.py:53(profile_init)
        1    0.000    0.000    4.229    4.229 {built-in method builtins.exec}
  1000000    0.044    0.000    0.044    0.000 {built-in method builtins.hasattr}
 13000000    0.487    0.000    0.487    0.000 {built-in method builtins.isinstance}
        1    0.000    0.000    0.000    0.000 {method 'disable' of '_lsprof.Profiler' objects}
  1000000    0.032    0.000    0.032    0.000 {method 'items' of 'dict' objects}
```

Benchmark code:
```python
import cProfile
from datetime import datetime
from typing import Any, Callable

import frappe
from frappe.model.base_document import BaseDocument

frappe.init("site", "../../sites")
frappe.connect()


def run_1_000_000(func: Callable[..., Any]):
    def inner(*args, **kwargs):
        for _ in range(1_000_000):
            func(*args, **kwargs)

    return inner


doc = frappe.new_doc("ToDo")
_update_data = {
    "status": "Closed",
    "priority": "High",
    "color": "blue",
    "date": datetime.now().date(),
    "owner": "Administrator",
    "description": "lalala",
    "reference_type": "DocType",
    "reference_name": "ToDo",
}


@run_1_000_000
def profile_update():
    # on develop: 2.692
    # on base-document-perf: 3.651
    doc.update(_update_data)


_document_dict = {
    "doctype": "ToDo",
    "docstatus": None,
    "status": "Closed",
    "priority": "High",
    "color": "blue",
    "date": datetime.now().date(),
    "owner": "Administrator",
    "description": "lalala",
    "reference_type": "DocType",
    "reference_name": "ToDo",
}


@run_1_000_000
def profile_init():
    BaseDocument(_document_dict)


# cProfile.runctx("profile_update()", globals(), locals())
cProfile.runctx("profile_init()", globals(), locals())

```