@gavindsouza 2021-12-28
@vrslev LGTM to me - the removal of the redundant line & using subscripting instead of get in `BaseDocument.update`